refactor(editor): move middleware config to separate module
